Refactor old signup tests
